+++ Spot.IM-Development/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- engine-monetization-view event [#120](https://github.com/SpotIM/spotim-ios-sdk-demo-apps/pull/120)
+
 ## [1.0.2] - 2020-05-19
 ### Added
 - New target type SpotImNoAds to support non-monetized partners (Fox)